Develop

+++ source/kernel/cpu/isr.cpp
@@ -78,41 +78,41 @@ namespace mirus
         {
             memset_v(isr_routines, 0x00, sizeof(isr_routines));
 
-            idt::set_gate(0, (unsigned)isr0, 0x08, 0xEE);
-            idt::set_gate(1, (unsigned)isr1, 0x08, 0x8E);
-            idt::set_gate(2, (unsigned)isr2, 0x08, 0x8E);
-            idt::set_gate(3, (unsigned)isr3, 0x08, 0x8E);
-            idt::set_gate(4, (unsigned)isr4, 0x08, 0x8E);
-            idt::set_gate(5, (unsigned)isr5, 0x08, 0x8E);
-            idt::set_gate(6, (unsigned)isr6, 0x08, 0x8E);
-            idt::set_gate(7, (unsigned)isr7, 0x08, 0x8E);
+            idt::set_gate(0, (unsigned long)isr0, 0x08, 0xEE);
+            idt::set_gate(1, (unsigned long)isr1, 0x08, 0x8E);
+            idt::set_gate(2, (unsigned long)isr2, 0x08, 0x8E);
+            idt::set_gate(3, (unsigned long)isr3, 0x08, 0x8E);
+            idt::set_gate(4, (unsigned long)isr4, 0x08, 0x8E);
+            idt::set_gate(5, (unsigned long)isr5, 0x08, 0x8E);
+            idt::set_gate(6, (unsigned long)isr6, 0x08, 0x8E);
+            idt::set_gate(7, (unsigned long)isr7, 0x08, 0x8E);
 
-            idt::set_gate(8, (unsigned)isr8, 0x08, 0x8E);
-            idt::set_gate(9, (unsigned)isr9, 0x08, 0x8E);
-            idt::set_gate(10, (unsigned)isr10, 0x08, 0x8E);
-            idt::set_gate(11, (unsigned)isr11, 0x08, 0x8E);
-            idt::set_gate(12, (unsigned)isr12, 0x08, 0x8E);
-            idt::set_gate(13, (unsigned)isr13, 0x08, 0x8E);
-            idt::set_gate(14, (unsigned)isr14, 0x08, 0x8E);
-            idt::set_gate(15, (unsigned)isr15, 0x08, 0x8E);
+            idt::set_gate(8, (unsigned long)isr8, 0x08, 0x8E);
+            idt::set_gate(9, (unsigned long)isr9, 0x08, 0x8E);
+            idt::set_gate(10, (unsigned long)isr10, 0x08, 0x8E);
+            idt::set_gate(11, (unsigned long)isr11, 0x08, 0x8E);
+            idt::set_gate(12, (unsigned long)isr12, 0x08, 0x8E);
+            idt::set_gate(13, (unsigned long)isr13, 0x08, 0x8E);
+            idt::set_gate(14, (unsigned long)isr14, 0x08, 0x8E);
+            idt::set_gate(15, (unsigned long)isr15, 0x08, 0x8E);
 
-            idt::set_gate(16, (unsigned)isr16, 0x08, 0x8E);
-            idt::set_gate(17, (unsigned)isr17, 0x08, 0x8E);
-            idt::set_gate(18, (unsigned)isr18, 0x08, 0x8E);
-            idt::set_gate(19, (unsigned)isr19, 0x08, 0x8E);
-            idt::set_gate(20, (unsigned)isr20, 0x08, 0x8E);
-            idt::set_gate(21, (unsigned)isr21, 0x08, 0x8E);
-            idt::set_gate(22, (unsigned)isr22, 0x08, 0x8E);
-            idt::set_gate(23, (unsigned)isr23, 0x08, 0x8E);
+            idt::set_gate(16, (unsigned long)isr16, 0x08, 0x8E);
+            idt::set_gate(17, (unsigned long)isr17, 0x08, 0x8E);
+            idt::set_gate(18, (unsigned long)isr18, 0x08, 0x8E);
+            idt::set_gate(19, (unsigned long)isr19, 0x08, 0x8E);
+            idt::set_gate(20, (unsigned long)isr20, 0x08, 0x8E);
+            idt::set_gate(21, (unsigned long)isr21, 0x08, 0x8E);
+            idt::set_gate(22, (unsigned long)isr22, 0x08, 0x8E);
+            idt::set_gate(23, (unsigned long)isr23, 0x08, 0x8E);
 
-            idt::set_gate(24, (unsigned)isr24, 0x08, 0x8E);
-            idt::set_gate(25, (unsigned)isr25, 0x08, 0x8E);
-            idt::set_gate(26, (unsigned)isr26, 0x08, 0x8E);
-            idt::set_gate(27, (unsigned)isr27, 0x08, 0x8E);
-            idt::set_gate(28, (unsigned)isr28, 0x08, 0x8E);
-            idt::set_gate(29, (unsigned)isr29, 0x08, 0x8E);
-            idt::set_gate(30, (unsigned)isr30, 0x08, 0x8E);
-            idt::set_gate(31, (unsigned)isr31, 0x08, 0x8E);
+            idt::set_gate(24, (unsigned long)isr24, 0x08, 0x8E);
+            idt::set_gate(25, (unsigned long)isr25, 0x08, 0x8E);
+            idt::set_gate(26, (unsigned long)isr26, 0x08, 0x8E);
+            idt::set_gate(27, (unsigned long)isr27, 0x08, 0x8E);
+            idt::set_gate(28, (unsigned long)isr28, 0x08, 0x8E);
+            idt::set_gate(29, (unsigned long)isr29, 0x08, 0x8E);
+            idt::set_gate(30, (unsigned long)isr30, 0x08, 0x8E);
+            idt::set_gate(31, (unsigned long)isr31, 0x08, 0x8E);
         }
 
         extern "C" void fault_handler(cpu::regs* r)
